Add a parameter that accounts for additional, transient memory used during servable load that is released when the loading phase is complete.
This parameter is a bit of a hack, among other things because it doesn't permit per-model differentiation. The long-term plan is to move resource estimation to an earlier stage than tensorflow-serving and store the estimates in the SavedModels (we'll have two estimates there: one for the during-load memory usage and one for the post-load usage). At that time we can remove this parameter.
Change: 165009205